chore: update bd

+++ prisma/migrations/20230422165556_update_db/migration.sql
@@ -15,7 +15,7 @@ CREATE TABLE "Product" (
     "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
     "description" TEXT NOT NULL,
-    "commission" DOUBLE PRECISION NOT NULL,
+    "commission" INTEGER NOT NULL,
     "score" INTEGER NOT NULL,
     "price" DOUBLE PRECISION NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
